use of verbs and states in api rest

+++ src/main/java/cl/people/jrios/controller/CourseController.java
@@ -1,8 +1,13 @@
 package cl.people.jrios.controller;
 
+import java.net.URI;
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,7 +16,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import cl.people.jrios.exception.ModelNotFoundException;
 import cl.people.jrios.model.entity.Course;
 import cl.people.jrios.service.ICourseService;
 
@@ -23,28 +30,43 @@ public class CourseController {
 	private ICourseService service;
 
 	@GetMapping
-	public List<Course> toList() {
-		return service.toList();
+	public ResponseEntity<List<Course>> toList() {
+		List<Course> list = service.toList();
+		return new ResponseEntity<List<Course>>(list, HttpStatus.OK);
 	}
 
 	@GetMapping("/{id}")
-	public Course toListForId(@PathVariable("id") Integer id) {
-		return service.readForId(id);
+	public ResponseEntity<Course> toListForId(@PathVariable("id") Integer id) {
+		Course cour = service.readForId(id);
+		// if it detects that the id does not exist, send exception
+		if (cour.getIdCourse() == null) {
+			throw new ModelNotFoundException("ID COURSE NOT FOUND " + id);
+		}
+		return new ResponseEntity<Course>(cour, HttpStatus.CREATED);
 	}
 
 	@PostMapping
-	public Course register(@RequestBody Course course) {
-		return service.register(course);
+	public ResponseEntity<Object> register(@Valid @RequestBody Course course) {
+		Course cour = service.register(course);
+		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
+				.buildAndExpand(course.getIdCourse()).toUri();
+		return ResponseEntity.created(location).build();
 	}
 
 	@PutMapping
-	public Course modify(@RequestBody Course course) {
-		return service.modify(course);
+	public ResponseEntity<Course> modify(@Valid @RequestBody Course course) {
+		Course cour = service.modify(course);
+		return new ResponseEntity<Course>(cour, HttpStatus.OK);
 	}
 
 	@DeleteMapping("/{id}")
-	public void remove(@PathVariable("id") Integer id) {
+	public ResponseEntity<Object> remove(@PathVariable("id") Integer id) {
+		Course cour = service.readForId(id);
+		if (cour.getIdCourse() == null) {
+			throw new ModelNotFoundException("ID COURSE NOT FOUND " + id);
+		}
 		service.remove(id);
+		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 
 }
