+++ src/main/java/cl/people/jrios/controller/StudentController.java
@@ -1,8 +1,13 @@
 package cl.people.jrios.controller;
 
+import java.net.URI;
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,7 +16,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
+import cl.people.jrios.exception.ModelNotFoundException;
 import cl.people.jrios.model.entity.Student;
 import cl.people.jrios.service.IStudentService;
 
@@ -23,28 +30,42 @@ public class StudentController {
 	private IStudentService service;
 
 	@GetMapping
-	public List<Student> toList() {
-		return service.toList();
+	public ResponseEntity<List<Student>> toList() {
+		List<Student> list = service.toList();
+		return new ResponseEntity<List<Student>>(list, HttpStatus.OK);
 	}
 
 	@GetMapping("/{id}")
-	public Student toListForId(@PathVariable("id") Integer id) {
-		return service.readForId(id);
+	public ResponseEntity<Student> toListForId(@PathVariable("id") Integer id) {
+		Student stu = service.readForId(id);
+		// if it detects that the id does not exist, send exception
+		if(stu.getIdStudent() == null) {
+			throw new ModelNotFoundException("ID STUDENT NOT FOUND " + id);
+		}
+		return new ResponseEntity<Student>(stu, HttpStatus.CREATED);
 	}
 
 	@PostMapping
-	public Student register(@RequestBody Student student) {
-		return service.register(student);
+	public ResponseEntity<Object> register(@Valid @RequestBody Student student) {
+		Student stu = service.register(student);
+		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(student.getIdStudent()).toUri();
+		return ResponseEntity.created(location).build();
 	}
 
 	@PutMapping
-	public Student modify(@RequestBody Student student) {
-		return service.modify(student);
+	public ResponseEntity<Student> modify(@Valid @RequestBody Student student) {
+		Student stu = service.modify(student);
+		return new ResponseEntity<Student>(stu, HttpStatus.OK);
 	}
 
 	@DeleteMapping("/{id}")
-	public void remove(@PathVariable("id") Integer id) {
+	public ResponseEntity<Object> remove(@PathVariable("id") Integer id) {
+		Student stu = service.readForId(id);
+		if(stu.getIdStudent() == null) {
+			throw new ModelNotFoundException("ID STUDENT NOT FOUND " + id);
+		}
 		service.remove(id);
+		return new ResponseEntity<Object>(HttpStatus.OK);
 	}
 
 }
